раскидал весь код по своим файлам
