+++ CSharp1Level/HomeWork1Console/Task5.cs
@@ -1,0 +1,17 @@
+﻿namespace HomeWork1Console
+{
+    public class Task5 : TaskWork
+    {
+        public void Work()
+        {
+            PrintMyInfo("Иван", "Бармин", "НЕ СКАЖУ!!!");
+        }
+        void PrintMyInfo(string firstName, string lastName, string city)
+        {
+            //Можно ещё уменьшить повторяемость кода с коллекциями, но это уже совсем другая история
+            Helper.WriteMsgInCenter("моё имя", firstName);
+            Helper.WriteMsgInCenter("фамилия", lastName);
+            Helper.WriteMsgInCenter("город проживания", city);
+        }
+    }
+}
